+++ redwood-protocol-guest/src/commonTest/kotlin/app/cash/redwood/protocol/guest/ProtocolTest.kt
@@ -16,16 +16,10 @@
 package app.cash.redwood.protocol.guest
 
 import androidx.compose.runtime.BroadcastFrameClock
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import app.cash.redwood.compose.WidgetVersion
-import app.cash.redwood.layout.compose.Column
-import app.cash.redwood.layout.compose.Row
-import app.cash.redwood.lazylayout.compose.ExperimentalRedwoodLazyLayoutApi
-import app.cash.redwood.lazylayout.compose.LazyColumn
 import app.cash.redwood.protocol.Change
 import app.cash.redwood.protocol.ChildrenChange
 import app.cash.redwood.protocol.ChildrenTag
@@ -43,12 +37,8 @@ import app.cash.redwood.ui.Cancellable
 import app.cash.redwood.ui.OnBackPressedCallback
 import app.cash.redwood.ui.OnBackPressedDispatcher
 import app.cash.redwood.ui.UiConfiguration
-import assertk.assertFailure
 import assertk.assertThat
-import assertk.assertions.containsExactly
 import assertk.assertions.isEqualTo
-import assertk.assertions.isInstanceOf
-import assertk.assertions.message
 import com.example.redwood.testapp.compose.Button
 import com.example.redwood.testapp.compose.Button2
 import com.example.redwood.testapp.compose.TestRow
@@ -298,97 +288,6 @@ class ProtocolTest {
     )
   }
 
-  @Test fun entireSubtreeRemovedLatest() = runTest {
-    assertThat(removeSubtree(latestVersion))
-      .containsExactly(
-        ChildrenChange.Remove(Id.Root, ChildrenTag.Root, 0, 1),
-      )
-  }
-
-  @Test fun entireSubtreeRemovedOldHostSynthesizesDepthFirstRemoval() = runTest {
-    assertThat(removeSubtree(RedwoodVersion("0.9.0")))
-      .containsExactly(
-        ChildrenChange.Remove(Id(2), ChildrenTag(1), 0, 1, listOf(Id(3))),
-        ChildrenChange.Remove(Id(1), ChildrenTag(1), 0, 1, listOf(Id(2))),
-        ChildrenChange.Remove(Id.Root, ChildrenTag.Root, 0, 1, listOf(Id(1))),
-      )
-  }
-
-  @Test fun entireSubtreeRemovedForLazyListPlaceholders() = runTest {
-    assertThat(removeSubtree(latestVersion, LazyListParent))
-      .containsExactly(
-        ChildrenChange.Remove(Id.Root, ChildrenTag.Root, 0, 1),
-      )
-  }
-
-  /**
-   * Our LazyList binding on host platforms incorrectly assumed that the placeholders children was
-   * append-only. When we fixed a host-side memory leak by traversing guest children, that
-   * introduced a crash. Special-case this by not synthesizing subtree removal for these children.
-   */
-  @Test fun entireSubtreeNotRemovedForLazyListPlaceholders() = runTest {
-    assertThat(removeSubtree(RedwoodVersion("0.9.0"), LazyListParent))
-      .containsExactly(
-        ChildrenChange.Remove(Id(2), ChildrenTag(2), 0, 1, listOf(Id(23))),
-        ChildrenChange.Remove(Id(1), ChildrenTag(1), 0, 1, listOf(Id(2))),
-        ChildrenChange.Remove(Id.Root, ChildrenTag.Root, 0, 1, listOf(Id(1))),
-      )
-  }
-
-  @Test fun entireSubtreeRemovedForRefreshableLazyListPlaceholders() = runTest {
-    assertThat(removeSubtree(latestVersion, RefreshableLazyListParent))
-      .containsExactly(
-        ChildrenChange.Remove(Id.Root, ChildrenTag.Root, 0, 1),
-      )
-  }
-
-  @Test fun entireSubtreeNotRemovedForRefreshableLazyListPlaceholders() = runTest {
-    assertThat(removeSubtree(RedwoodVersion("0.9.0"), RefreshableLazyListParent))
-      .containsExactly(
-        ChildrenChange.Remove(Id(2), ChildrenTag(2), 0, 1, listOf(Id(23))),
-        ChildrenChange.Remove(Id(1), ChildrenTag(1), 0, 1, listOf(Id(2))),
-        ChildrenChange.Remove(Id.Root, ChildrenTag.Root, 0, 1, listOf(Id(1))),
-      )
-  }
-
-  private suspend fun TestScope.removeSubtree(
-    hostVersion: RedwoodVersion,
-    parent: SubtreeParent = ColumnParent,
-  ): List<Change> {
-    val (composition, guest) = testProtocolComposition(hostVersion)
-
-    var clicks = 0
-    var remove by mutableStateOf(false)
-    composition.setContent {
-      if (!remove) {
-        Row {
-          parent.Wrap {
-            Button("Click?", onClick = { clicks++ })
-          }
-        }
-      }
-    }
-    val initialSnapshot = composition.awaitSnapshot()
-    val button = initialSnapshot.first { it is Create && it.tag.value == 4 }
-    assertThat(clicks).isEqualTo(0)
-
-    // Ensure the button is present and receiving clicks.
-    guest.sendEvent(Event(button.id, EventTag(2)))
-    assertThat(clicks).isEqualTo(1)
-
-    remove = true
-    val removeChanges = composition.awaitSnapshot()
-
-    // If the whole tree was removed, we cannot target the button anymore.
-    assertFailure { guest.sendEvent(Event(button.id, EventTag(2))) }
-      .isInstanceOf<IllegalArgumentException>()
-      .message()
-      .isEqualTo("Unknown node ID ${button.id.value} for event with tag 2")
-    assertThat(clicks).isEqualTo(1)
-
-    return removeChanges
-  }
-
   private fun TestScope.testProtocolComposition(
     hostVersion: RedwoodVersion = latestVersion,
   ): Pair<TestRedwoodComposition<List<Change>>, GuestProtocolAdapter> {
@@ -407,53 +306,5 @@ class ProtocolTest {
       composition.cancel()
     }
     return composition to guestAdapter
-  }
-
-  interface SubtreeParent {
-    @Composable
-    fun Wrap(content: @Composable () -> Unit)
-  }
-
-  object ColumnParent : SubtreeParent {
-    @Composable
-    override fun Wrap(content: @Composable () -> Unit) {
-      Column {
-        content()
-      }
-    }
-  }
-
-  object LazyListParent : SubtreeParent {
-    @Composable
-    override fun Wrap(content: @Composable () -> Unit) {
-      LazyColumn(
-        placeholder = {
-          Text("placeholder")
-        },
-      ) {
-        item {
-          content()
-        }
-      }
-    }
-  }
-
-  object RefreshableLazyListParent : SubtreeParent {
-    @OptIn(ExperimentalRedwoodLazyLayoutApi::class)
-    @Composable
-    override fun Wrap(content: @Composable () -> Unit) {
-      LazyColumn(
-        refreshing = false,
-        onRefresh = {
-        },
-        placeholder = {
-          Text("placeholder")
-        },
-      ) {
-        item {
-          content()
-        }
-      }
-    }
   }
 }